wip(spike): consolidate connector compile pipelines
Three packages each shipped their own near-identical copy of
`findBundledConnectorFile` + `compileConnectorFromFile` + the `npm:`
specifier esbuild plugin + the `lobu`/`@lobu/connector-sdk` aliasing +
the mtime-keyed LRU compile cache:

  - packages/connector-worker/src/compile-connector.ts  (~150 LOC)
  - packages/cli/src/commands/_lib/connector-loader.ts  (~120 LOC)
  - packages/server/src/utils/connector-catalog.ts      (~150 LOC of the
                                                          file, the rest
                                                          stays — metadata
                                                          extraction, URI
                                                          normalisation,
                                                          catalog iteration)

Comments in two of the three already said things like "single source of
truth lives in @lobu/connector-worker. Duplicated here as a flat string
array to keep this file self-contained" and "Mirror the worker-side
resolver in compile-connector.ts: try the subdir layout first … Keep
these in sync if either side changes." That's the drift surface this
spike closes.

New shared module: `packages/connector-worker/src/compile/index.ts`,
exported as `@lobu/connector-worker/compile`. (Originally `./build` but
the repo root `.gitignore` has `**/build/` which silently ate the new
directory; renamed to `./compile` so it tracks.)

Surface:

  - `EXTERNAL_RUNTIME_DEPS` — re-exported from `runtime-deps.ts`.
  - `findBundledConnectorFile(key, candidateDirs)` — pure helper; the
    per-environment candidate-dir list is the parameter that genuinely
    differs (worker image vs CLI dist-bundle vs gateway pod). No
    internal cache; callers that need one (`server/connector-catalog.ts`,
    `cli/connector-loader.ts`) wrap their own thin Map<key, path|null>.
  - `createNpmSpecifierPlugin({ onUnresolved })` — esbuild plugin
    parametrised on the warn hook (server wants to log via pino, CLI/
    worker stay silent).
  - `createConnectorCompiler({ cacheMax?, sdkEntry?, onUnresolvedNpm? })`
    → `{ compileConnectorFromFile }`. Each instance owns its own
    8-entry LRU; callers that want per-process isolation create their
    own.

Net diff: -381 / +296 (with the new 229-LOC shared module) = -85 LOC
across the three packages plus zero risk of the three implementations
drifting from each other.

Server now declares `@lobu/connector-worker: workspace:*` as a real dep
(it was already importing from it after the prior round-2 spike, but
without a corresponding package.json entry).

Typecheck status (`make typecheck` from worktree root): **clean.** The
pre-existing `gateway/guardrails/aggregator.ts` `SkillPreToolGuardrail`
errors that REPORT.md called out turned out to be stale dist artifacts
in this worktree — once `make build-packages` re-emitted `@lobu/core`'s
.d.ts, those resolved too. Whole repo typecheck is green.
