+++ packages/server/package.json
@@ -37,6 +37,7 @@
     "@hono/zod-openapi": "^1.2.1",
     "@lobu/core": "workspace:*",
     "@lobu/connector-sdk": "workspace:*",
+    "@lobu/connector-worker": "workspace:*",
     "@mariozechner/pi-ai": "^0.51.6",
     "@modelcontextprotocol/sdk": "^1.27.1",
     "@polyglot-sql/sdk": "^0.1.13",
